+++ void_getting_start.sh
@@ -3,11 +3,11 @@
 #
 # --- Script Version ---
 # Name    : void_getting_start.sh
-# Version : 71f746f (1 commit after this ref)
+# Version : 39f2a0a (1 commit after this ref)
 # Author  : IceCodeNew
 # Date    : Wed Oct 20th, 2021
 # Download: https://cdn.jsdelivr.net/gh/IceCodeNew/go-collection@master/void_getting_start.sh
-readonly local_script_version='71f746f'
+readonly local_script_version='39f2a0a'
 
 curl_path="$(type -P curl)"
 # geo_country="$(curl 'https://api.myip.la/en?json' | jq . | grep country_code | cut -d'"' -f4)"
@@ -298,22 +298,24 @@ END_TEXT
       sudo rm -f /usr/bin/haproxy &&
       sudo ln -s /usr/local/sbin/haproxy /usr/bin/
 
-  sudo xbps-install -Su &&
-  sudo xbps-install -Su &&
-  sudo xbps-install -y caddy &&
-  ln -s /etc/sv/caddy /var/service/;
-  if [[ x"$(echo "${donot_need_caddy_autorun:=no}" | cut -c1)" = x'y' ]]; then
-    sudo touch /etc/sv/caddy/down
-  else
-    sudo sed -i -E 's/^:80/:19600/' /etc/caddy/Caddyfile
+  tmp_dir=$(mktemp -d)
+  pushd "$tmp_dir" || exit 1
+  if [[ ! -f /etc/caddy/Caddyfile ]]; then
+    sudo xbps-install -y caddy &&
+    ln -s /etc/sv/caddy /var/service/
+    if [[ x"$(echo "${donot_need_caddy_autorun:=no}" | cut -c1)" = x'y' ]]; then
+      sudo touch /etc/sv/caddy/down
+    else
+      sudo sed -i -E 's/^:80/:19600/' /etc/caddy/Caddyfile
+    fi
   fi
-  sudo rm '/usr/local/bin/caddy' '/usr/local/bin/xcaddy'
 
-  curl -L "https://cdn.jsdelivr.net/gh/IceCodeNew/go-collection@latest-release/assets/caddy.zst" | unzstd -q --no-progress -o './caddy' &&
-    sudo "$(type -P install)" -pvD './caddy' '/usr/local/sbin/caddy' &&
-  # curl_to_dest "https://github.com/IceCodeNew/go-collection/raw/latest-release/assets/caddy" '/usr/local/sbin/caddy' &&
-    sudo rm -f /usr/bin/caddy &&
-    sudo ln -s /usr/local/sbin/caddy /usr/bin/
+  sudo rm -f '/usr/local/bin/caddy' '/usr/local/bin/xcaddy'
+  curl -L "https://cdn.jsdelivr.net/gh/IceCodeNew/go-collection@latest-release/assets/caddy.zst" |
+    unzstd -q --no-progress -o './caddy' && sudo "$(type -P install)" -pvD './caddy' '/usr/local/sbin/caddy' &&
+    sudo rm -f '/usr/bin/caddy' && sudo ln -s /usr/local/sbin/caddy /usr/bin/
+  popd || exit 1
+  /bin/rm -rf "$tmp_dir"
 
   # tmp_dir=$(mktemp -d)
   # pushd "$tmp_dir" || exit 1
